# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a documentation for the Swipe API",
        "title": "Swipe Public API v0.0.1",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/purchase/carts": {
            "get": {
                "description": "get list of items from carts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchase"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.CartSlices"
                        }
                    }
                }
            },
            "post": {
                "description": "add item to carts.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchase"
                ],
                "parameters": [
                    {
                        "description": "cart insert request",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CartInsert"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.OK"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete item from carts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchase"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "inventories id",
                        "name": "wid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.OK"
                        }
                    }
                }
            }
        },
        "/purchase/orders": {
            "get": {
                "description": "get list of orders.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchase"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.OrderSchema"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "create order.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchase"
                ],
                "parameters": [
                    {
                        "description": "order insert request",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateOrderSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.OK"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.CartInsert": {
            "type": "object",
            "required": [
                "inventory_id",
                "quantity",
                "user_id"
            ],
            "properties": {
                "inventory_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "dtos.CartSchema": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "img": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "dtos.CartSlices": {
            "type": "object",
            "properties": {
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.CartSchema"
                    }
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "dtos.CreateOrderSchema": {
            "type": "object",
            "properties": {
                "product": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "inventory_id": {
                                "type": "integer"
                            },
                            "quantity": {
                                "type": "integer"
                            }
                        }
                    }
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "dtos.OK": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
        "dtos.OrderSchema": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.ProductOrderSchema"
                    }
                },
                "status": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "user_email": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "user_name": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "dtos.ProductOrderSchema": {
            "type": "object",
            "properties": {
                "currency_code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "inventory_id": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "total_amount": {
                    "type": "string"
                }
            }
        },
...
    }
}